move ColorSpec to forest.colors to ease usage
